chore: test with hds_console (#1)
This repository is a copy (not connected as a fork) of the
tokio-rs/console repo. It will be used to test release automation.

The first step is to rename all the crates to something different (in
this case, prefixing with `hds_`) so that they can be released without
affecting the existing tokio console ones.

Many tests and examples needed to be updated to point to the new crate
names (especially for the now named hds_console-subscriber).